Refactor existing in-memory timed caches to use TimeoutDict
