+++ src/vumi2/transports/smpp/smpp_cache.py
@@ -1,8 +1,9 @@
-from datetime import datetime
 from typing import Optional
 
 import cattrs
 from attrs import define
+
+from vumi2.message_caches import TimeoutDict
 
 
 class BaseSmppCache:  # pragma: no cover
@@ -35,7 +36,7 @@ class InMemorySmppCache(BaseSmppCache):
     def __init__(self, config: dict) -> None:
         self.config = cattrs.structure(config, InMemorySmppCacheConfig)
         self._multipart: dict[tuple[int, int], dict[int, str]] = {}
-        self._smpp_msg_id: dict[str, tuple[str, datetime]] = {}
+        self._smpp_msg_id: TimeoutDict[str] = TimeoutDict(self.config.timeout)
 
     async def store_multipart(
         self, ref_num: int, tot_num: int, part_num: int, content: str
@@ -53,27 +54,13 @@ class InMemorySmppCache(BaseSmppCache):
             return "".join(c for i, c in sorted(parts.items()))
         return None
 
-    async def _remove_expired(self):
-        now = datetime.now()
-        to_remove = []
-        for key, (_, timestamp) in self._smpp_msg_id.items():
-            if (now - timestamp).total_seconds() >= self.config.timeout:
-                to_remove.append(key)
-            else:
-                # Dictionaries are ordered, so we don't need to check all the others
-                # if we've found a new enough message
-                break
-        for key in to_remove:
-            await self.delete_smpp_message_id(key)
-
     async def store_smpp_message_id(
         self, vumi_message_id: str, smpp_message_id: str
     ) -> None:
         """
         Stores the mapping between one or many smpp message IDs and the vumi message ID
         """
-        await self._remove_expired()
-        self._smpp_msg_id[smpp_message_id] = (vumi_message_id, datetime.now())
+        self._smpp_msg_id[smpp_message_id] = vumi_message_id
 
     async def delete_smpp_message_id(self, smpp_message_id: str) -> None:
         """
@@ -85,5 +72,4 @@ class InMemorySmppCache(BaseSmppCache):
         """
         Returns the vumi message ID for the given smpp message id
         """
-        vumi_message_id, _ = self._smpp_msg_id.pop(smpp_message_id, (None, None))
-        return vumi_message_id
+        return self._smpp_msg_id.pop(smpp_message_id)
