+++ tests/transports/smpp/test_smpp_cache.py
@@ -1,6 +1,5 @@
-from datetime import datetime, timedelta
-
 import pytest
+import trio
 
 from vumi2.transports.smpp.smpp_cache import InMemorySmppCache
 
@@ -40,17 +39,19 @@ async def test_in_memory_delete_smpp_message_id(memory_smpp_cache: InMemorySmppC
     await memory_smpp_cache.delete_smpp_message_id("invalid")
 
 
-async def test_in_memory_remove_expired(memory_smpp_cache: InMemorySmppCache):
+async def test_in_memory_remove_expired(memory_smpp_cache, autojump_clock):
     """
     If items are old, they should be removed from the cache
     """
     await memory_smpp_cache.store_smpp_message_id("vumi2", "delete")
+    await trio.sleep(memory_smpp_cache.config.timeout - 1)
     await memory_smpp_cache.store_smpp_message_id("vumi1", "keep")
-    memory_smpp_cache._smpp_msg_id["delete"] = (
-        "vumi2",
-        datetime.now() - timedelta(hours=25),
-    )
-    await memory_smpp_cache._remove_expired()
 
+    # Neither entry has expired yet, although "delete" is close.
+    assert "delete" in memory_smpp_cache._smpp_msg_id
     assert "keep" in memory_smpp_cache._smpp_msg_id
-    assert "delete" not in memory_smpp_cache._smpp_msg_id
+
+    # Wait until "delete" expires, leaving only "keep".
+    await trio.sleep(2)
+    assert await memory_smpp_cache.get_smpp_message_id("delete") is None
+    assert await memory_smpp_cache.get_smpp_message_id("keep") == "vumi1"
